requirements.txt
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.0.3
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-language==2.17.0
google-cloud-speech==2.27.0
google-cloud-texttospeech==2.17.2
googleapis-common-protos==1.67.0
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==22.0.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.2.3
packaging==24.2
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
requests==2.32.3
rsa==4.9
urllib3==2.3.0
Werkzeug==3.1.3
